backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
httpx==0.28.1
python-dotenv==1.0.1
jinja2==3.1.4
aiofiles==24.1.0
